Revert "[http] Make chunk_remaining also off_t"
This reverts commit 7eb8fa563d51f77cbca1b9dcf458ad91f46a917c.

It was overkill.

+++ src/protocol/http/http.h
@@ -30,7 +30,7 @@ struct http_connection_info {
 
 	int close;
 	off_t length;
-	off_t chunk_remaining;
+	int chunk_remaining;
 	int code;
 
 	struct http_post post;
